+++ php/Terminus/Commands/TerminusCommand.php
@@ -176,14 +176,26 @@ abstract class TerminusCommand {
    * Outputs basic workflow success/failure messages
    *
    * @param Workflow $workflow Workflow to output message about
+   * @param array    $messages Messages to override workflow's defaults:
+   *  string success Success message to override workflow default
+   *  string failure Failure message to override workflow default
    * @return void
    */
-  protected function workflowOutput($workflow) {
+  protected function workflowOutput($workflow, array $messages = []) {
     if ($workflow->get('result') == 'succeeded') {
-      $this->log()->info($workflow->get('active_description'));
+      $message = $workflow->get('active_description');
+      if (isset($messages['success'])) {
+        $message = $messages['success'];
+      }
+      $this->log()->info($message);
     } else {
-      $final_task = $workflow->get('final_task');
-      $this->log()->error($final_task->reason);
+      $message = 'Workflow failed.';
+      if (isset($messages['failure'])) {
+        $message = $messages['failure'];
+      } elseif (!is_null($final_task = $workflow->get('final_task'))) {
+        $message = $final_task->reason;
+      }
+      $this->log()->error($message);
     }
   }